% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_spatial_functions.R
\name{beta.array}
\alias{beta.array}
\title{Compute beta diversity on arrays}
\usage{
beta.array(x, index.family = "sorensen", tree = F, fwCC)
}
\arguments{
\item{x}{numeric array of with 3 dimensions}

\item{index.family}{ family of dissimilarity indices, partial match of \code{"sorensen"} or \code{"jaccard"}.}

\item{tree}{ a phylogenetic tree of class phylo with tips names identic to species names from the community matrix.}

\item{fwCC}{vector of observations to be kept. It is a matrix returned
from raster::focalWeight and coerced to vector with as.vector()}
}
\description{
This function will compute beta diversity on arrays where
the first two dimensions are sites and the third dimension is species
}
