% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_spatial_functions.R
\name{y_mean.numeric}
\alias{y_mean.numeric}
\title{Compute avg values of y for numeric class}
\usage{
\method{y_mean}{numeric}(y)
}
\description{
Compute avg values of y for numeric class
}
\keyword{internal}
