% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_spatial_functions.R
\name{y_mean.Raster}
\alias{y_mean.Raster}
\title{Compute avg values of y for Raster class}
\usage{
\method{y_mean}{Raster}(y)
}
\description{
Compute avg values of y for Raster class
}
\keyword{internal}
