% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_spatial_functions.R
\name{y_mean}
\alias{y_mean}
\title{Compute avg values of y}
\usage{
y_mean(y)
}
\arguments{
\item{y}{Numeric or Extent object, or any object from which an
Extent object can be extracted. See \code{\link[terra]{crop}} for details.}
}
\description{
Compute avg values of y
}
\keyword{internal}
