% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_spatial_functions.R
\name{y_mean.SpatRaster}
\alias{y_mean.SpatRaster}
\title{Compute avg values of y for SpatRaster class}
\usage{
\method{y_mean}{SpatRaster}(y)
}
\description{
Compute avg values of y for SpatRaster class
}
\keyword{internal}
