% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_spatial_functions.R
\name{km2decdeg}
\alias{km2decdeg}
\title{Convert distance from kilometers to decimal degrees}
\usage{
km2decdeg(dist, y)
}
\arguments{
\item{dist}{Numeric. Vector of distances to be converted from km to
decimal degrees}

\item{y}{Numeric or Extent object, or any object from which an
Extent object can be extracted. See \code{\link[terra]{crop}} for details.
Latitude of location where degrees will be converted to kilometers}
}
\description{
Convert distance from kilometers to decimal degrees
}
